Displays credit screen
Implement missing OAM-DMA transfer - used by tetris (and most
games) to quickly transfer objects to OAM.

Joypad register lower nibble was not read-only. So software was able to
overwrite it to 0 (which is all buttons pressed) which tetris considers
a soft reset.

VBLANK interrupt was triggering at the start of every VBLANK scanline
instead of just at LY=144