Fixed credit screen in release build bug
The issue was not related to undefined behavior / compiler optimizations
Instead when the boot rom ran, the credits screen would not display.
Boot rom was running after a zeroed out memory. This was wrong.
P1 joypad reg should cold reset to FF. This denotes "no buttons
pressed". My code to preserve the lower nibble bits only worked when it
was manually assigned to the right value (which only happened when we
skipped boot rom"